hackathon/requirements.txt
streamlit
fastapi
uvicorn
transformers
torch
sentence-transformers
pillow
torchvision
